- Change all BIT_{SET,CLR,TEST} to EBIT_ versions with enum types - Change CWCB callback to use size_t size instead of int. - Added LOG_TCP_DENIED

+++ src/pass.c
@@ -288,7 +288,7 @@ passWriteClient(int fd, void *data)
 }
 
 static void
-passErrorComplete(int fdnotused, void *passState, int sizenotused)
+passErrorComplete(int fdnotused, void *passState, size_t sizenotused)
 {
     assert(passState != NULL);
     passClose(passState);
